backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0

# Azure OpenAI - Support both approaches
azure-ai-inference>=1.0.0
azure-identity>=1.15.0

# Utils
python-multipart==0.0.6 